+++ .env.example
@@ -12,15 +12,13 @@ GOOGLE_GENAI_USE_VERTEXAI="False"
 
 # --- Service URLs ---
 # The URL where the Specialist Agent's A2A server will run
-STOCK_INFO_AGENT_A2A_SERVER_HOST=127.0.0.1
-STOCK_INFO_AGENT_A2A_SERVER_PORT=8001
-STOCK_INFO_AGENT_MODEL="gemini-2.0-flash-001"
+BLOCKCHAIN_INFO_AGENT_A2A_SERVER_HOST=127.0.0.1
+BLOCKCHAIN_INFO_AGENT_A2A_SERVER_PORT=8001
+BLOCKCHAIN_INFO_AGENT_MODEL="gemini-2.0-flash-001"
 MOCK_STOCK_API=False
-FINNHUB_API_KEY=<YOUR_FINNHUB_API_KEY>
-
 
 # --- MCP Server Configuration ---
-STOCK_MCP_SERVER_PATH="mcp_servers/stock_mcp_server/server.py" # Relative path from specialist agent perspective
+BLOCKCHAIN_MCP_SERVER_PATH="mcp_servers/blockchain_mcp_server/server.py" # Relative path from specialist agent perspective
 
 # --- ADK Live Server Configuration ---
 # Host and port for the main ADK Live Server (FastAPI/WebSocket server that serves the UI).
@@ -28,9 +26,9 @@ LIVE_SERVER_HOST=127.0.0.1
 LIVE_SERVER_PORT=8000 # MANDATORY: Port for the ADK Live Server (e.g., 8000). The UI will connect here.
 LIVE_SERVER_MODEL="gemini-2.0-flash-live-001"
 
-# Base URL for the StockInfoAgent (A2A Server). The Host Agent will append /.well-known/agent.json to this.
+# Base URL for the BlockchainInfoAgent (A2A Server). The Host Agent will append /.well-known/agent.json to this.
 # If you have multiple, comma-separate them: "http://localhost:8001,http://localhost:8002"
-SPECIALIST_AGENT_BASE_URLS="http://localhost:8001" # Points to StockInfoAgent A2A server
+SPECIALIST_AGENT_BASE_URLS="http://localhost:8001" # Points to BlockchainInfoAgent A2A server
 
 # Optional: Logging Level (DEBUG, INFO, WARNING, ERROR, CRITICAL)
 LOG_LEVEL="INFO"
