+++ examples/blockchain_info_lookup_a2a/app/static/utils/utils.js
